feat: Unified sdk set up
