+++ samples/kotlin-android-app/src/unified/AndroidManifest.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <application
+        android:name=".UnifiedApp"
+        android:icon="@mipmap/ic_launcher"
+        tools:replace="android:name" />
+
+</manifest>
