+++ samples/kotlin-android-app/src/unified/java/com/amplitude/android/sample/UnifiedApp.kt
@@ -1,0 +1,28 @@
+package com.amplitude.android.sample
+
+import com.amplitude.android.Configuration
+import com.amplitude.android.autocaptureOptions
+import com.amplitude.android.unified.Amplitude
+
+class UnifiedApp : MainApplication() {
+    override fun initAmplitudeLibrary() {
+        val httpClient = CustomOkHttpClient()
+        val configuration =
+            Configuration(
+                apiKey = AMPLITUDE_API_KEY,
+                context = applicationContext,
+                autocapture =
+                    autocaptureOptions {
+                        +sessions
+                        +appLifecycles
+                        +deepLinks
+                        +screenViews
+                        +elementInteractions
+                    },
+                httpClient = httpClient,
+                minTimeBetweenSessionsMillis = 60_000L,
+            )
+        httpClient.initialize(configuration)
+        amplitude = Amplitude(configuration)
+    }
+}
